+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+- The action detail of a chunked DEPLOY reports which action deployed the contract and whether the assembly is still pending.
+
 ## [0.15.2] - 2026-09-07
 
 ### Fixed